+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 0.1.0+1 - 2025-03-24
 - Initial release of Camera With GPS plugin.
 - Supports capturing photos and embedding GPS metadata on both iOS and Android.
+## 0.1.1+1 - 2025-03-31
+ - Update `openCamera` to require `BuildContext` for navigation 
+ - Replaced `navigatorKey` with a `BuildContext` parameter in `openCamera` for improved navigation handling.